removing boilerplate from myshop
