+++ {{cookiecutter.project_slug}}/{{cookiecutter.app_name}}/serializers.py
@@ -2,25 +2,18 @@
 from __future__ import unicode_literals
 
 from django.utils.safestring import mark_safe
-
 from rest_framework import serializers
-
-from shop.serializers.bases import ProductSerializer
 from shop.search.serializers import ProductSearchSerializer as BaseProductSearchSerializer
+from shop.serializers.bases import ProductSerializer
+from shop.serializers.defaults import AddToCartSerializer
 
+from {{ cookiecutter.app_name }} import SmartCard, SmartPhoneModel
 from {{ cookiecutter.app_name }}.search_indexes import myshop_search_index_classes
-
-__all__ = ['ProductSummarySerializer', 'ProductSearchSerializer', 'CatalogSearchSerializer']
 
 
 class ProductSummarySerializer(ProductSerializer):
-    media = serializers.SerializerMethodField()
-
     class Meta(ProductSerializer.Meta):
         fields = ['id', 'product_name', 'product_url', 'product_model', 'price', 'media', 'caption']
-
-    def get_media(self, product):
-        return self.render_html(product, 'media')
 
 {% if cookiecutter.products_model == 'commodity' %}
 
@@ -38,14 +31,6 @@ class ProductDetailSerializer(ProductSerializer):
                   'product_code', 'storage']
 
 __all__.append('ProductDetailSerializer')
-
-{% elif cookiecutter.products_model == 'polymorphic' %}
-
-from .polymorphic import (SmartCardSerializer, SmartPhoneSerializer, AddSmartPhoneToCartSerializer)
-
-__all__.extend(['SmartCardSerializer', 'SmartPhoneSerializer', 'AddSmartPhoneToCartSerializer'])
-
-{% endif %}
 
 class ProductSearchSerializer(BaseProductSearchSerializer):
     """
@@ -77,3 +62,41 @@ class CatalogSearchSerializer(BaseProductSearchSerializer):
 
     def get_media(self, search_result):
         return mark_safe(search_result.catalog_media)
+
+
+class SmartCardSerializer(ProductSerializer):
+    class Meta:
+        model = SmartCard
+        fields = ['product_name', 'slug', 'unit_price', 'manufacturer', 'card_type', 'speed',
+                  'product_code', 'storage']
+
+class SmartPhoneSerializer(ProductSerializer):
+    class Meta:
+        model = SmartPhoneModel
+        fields = ['product_name', 'slug', 'battery_type', 'battery_capacity']
+
+class AddSmartPhoneToCartSerializer(AddToCartSerializer):
+    """
+    Modified AddToCartSerializer which handles SmartPhones
+    """
+    def get_instance(self, context, data, extra_args):
+        product = context['product']
+        if data is empty:
+            product_code = None
+            extra = {}
+        else:
+            product_code = data.get('product_code')
+            extra = data.get('extra', {})
+        try:
+            variant = product.get_product_variant(product_code=product_code)
+        except product.DoesNotExist:
+            variant = product.variants.first()
+        extra.update(storage=variant.storage)
+        instance = {
+            'product': product.id,
+            'product_code': variant.product_code,
+            'unit_price': variant.unit_price,
+            'extra': extra,
+        }
+        return instance
+
